requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
botocore==1.21.65
bs4==0.0.1
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-crontab==0.7.1
django-ipware==4.0.0
djangorestframework==3.13.1
dnspython==2.1.0
dnstwist==20220120
future==0.16.0
geoip2==0.1.0
gunicorn==20.1.0
Jinja2==3.0.3
jmespath==0.10.01
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.9
pathspec==0.5.9
Pillow==8.3.2
ppdeep==20200505
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pylibscrypt==2.0.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-magic==0.4.26
pytz==2021.3
PyYAML==5.4.1
querycontacts==2.0.0
requests==2.26.0
scrypt==0.8.19
simplejson==3.17.6
six==1.14.0
soupsieve==2.3.1
sqlparse==0.4.2
stix2==3.0.1
stix2-patterns==1.3.2
urllib3==1.26.7
validate-email-address==1
voluptuous==0.12.2
w3lib==1.22.0
wcwidth==0.1.9
whitenoise==5.3.0
whois==0.9.13
xlrd==2.0.1
xlwt==1.3.0
